refactor: dango remote context into context
