+++ ui/store/src/remote.ts
@@ -1,10 +1,5 @@
 "use client";
 
-import { createElement } from "react";
-import { Hydrate } from "./hydrate.js";
-import { createConfig } from "./createConfig.js";
-import { graphql } from "@left-curve/dango";
-import { DangoStoreContext } from "./context.js";
 import { uid } from "@left-curve/dango/utils";
 import { deserializeJson, serializeJson } from "@left-curve/dango/encoding";
 
@@ -23,26 +18,6 @@ export interface WindowDangoStore extends Window {
 }
 
 declare let window: WindowDangoStore;
-
-export const DangoRemoteProvider: React.FC<React.PropsWithChildren> = (parameters) => {
-  const { children } = parameters;
-
-  const chain = window.dango.chain;
-
-  const config = createConfig({
-    chain,
-    transport: graphql(chain.urls.indexer, { batch: true }),
-    coins: window.dango.coins,
-    ssr: false,
-    connectors: [],
-  });
-
-  return createElement(
-    Hydrate,
-    { config, reconnectOnMount: false },
-    createElement(DangoStoreContext.Provider, { value: config }, children),
-  );
-};
 
 export const requestRemote = async <T = unknown>(
   method: string,
